refactor: 지원예정 혹은 서류 전형으로 지원서를 작성할시, 지원예정, 서류 전형 둘다 할당하는 로직 Application에서 Processes로 이전
