+++ core-domain/src/test/java/gohigher/common/ProcessesTest.java
@@ -1,15 +1,22 @@
 package gohigher.common;
 
 import static gohigher.application.ProcessFixture.*;
+import static gohigher.common.ProcessType.DOCUMENT;
+import static gohigher.common.ProcessType.TO_APPLY;
 import static org.assertj.core.api.Assertions.*;
 import static org.junit.jupiter.api.Assertions.*;
 
+import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.EnumSource;
+
+import gohigher.application.ProcessFixture;
 
 @DisplayName("Processes 클래스의")
 class ProcessesTest {
@@ -22,8 +29,8 @@ class ProcessesTest {
 		@Nested
 		class Context_with_processes_that_not_assigned_order {
 
-			private final Process toApply = TO_APPLY.toDomainWithoutOrder();
-			private final Process document = DOCUMENT.toDomainWithoutOrder();
+			private final Process toApply = ProcessFixture.TO_APPLY.toDomainWithoutOrder();
+			private final Process document = ProcessFixture.DOCUMENT.toDomainWithoutOrder();
 			private final Process test = TEST.toDomainWithoutOrder();
 			private final Process interview = INTERVIEW.toDomainWithoutOrder();
 			private final List<Process> input = List.of(toApply, document, test, interview);
@@ -56,6 +63,50 @@ class ProcessesTest {
 				assertAll(
 					() -> assertThat(processes.getValues()).hasSize(1),
 					() -> assertThat(processes.getValues().get(0).getType()).isEqualTo(ProcessType.TO_APPLY)
+				);
+			}
+		}
+
+		@DisplayName("1개의 지원 예정 혹은 서류 전형을 받을 경우")
+		@Nested
+		class Context_with_Only_Process_Of_To_Apply_Or_Document {
+
+			@DisplayName("동일한 일자의 지원 예정, 서류 전형을 가진 Processes를 반환한다.")
+			@ParameterizedTest
+			@EnumSource(value = ProcessType.class, names = {"TO_APPLY", "DOCUMENT"})
+			void it_returns_processes_with_to_apply_and_document(ProcessType processType) {
+				Process process = new Process(null, processType, "세부직무", LocalDateTime.now());
+
+				Processes actual = Processes.initialFrom(process);
+
+				List<Process> processes = actual.getValues();
+				Process firstProcess = processes.get(0);
+				Process secondProcess = processes.get(1);
+				List<ProcessType> processesTypes = List.of(firstProcess.getType(), secondProcess.getType());
+				assertAll(
+					() -> assertThat(processes).hasSize(2),
+					() -> assertThat(firstProcess.getSchedule()).isEqualTo(secondProcess.getSchedule()),
+					() -> assertThat(processesTypes).containsExactly(TO_APPLY, DOCUMENT)
+				);
+			}
+		}
+
+		@DisplayName("지원 예정 혹은 서류가 아닌 1개의 전형을 받을 경우")
+		@Nested
+		class Context_with_Only_Process_That_Is_Not_To_Apply_Or_Document {
+
+			@DisplayName("해당 전형만 가진 Processes를 반환한다.")
+			@ParameterizedTest
+			@EnumSource(value = ProcessType.class, names = {"TEST", "INTERVIEW", "COMPLETE"})
+			void it_returns_processes_with_only_that_process(ProcessType processType) {
+				Process process = new Process(null, processType, "세부직무", LocalDateTime.now());
+
+				Processes actual = Processes.initialFrom(process);
+
+				List<Process> processes = actual.getValues();
+				assertAll(
+					() -> assertThat(processes).hasSize(1),
+					() -> assertThat(processes.get(0).getType()).isEqualTo(processType)
 				);
 			}
 		}
